requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
psutil==5.9.8
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.3
SQLAlchemy==2.0.25
alembic==1.13.1
python-dateutil==2.8.2
pytz==2024.1
urllib3==2.1.0
cryptography==42.0.2
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2
MarkupSafe==2.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
typing_extensions==4.9.0
Mako==1.3.2
colorama==0.4.6
cffi==1.16.0
pycparser==2.21
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
PyJWT==2.9.0
celery==5.4.0
redis==5.0.8
